% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdh.aurisano.check.plot.R
\name{bmdh.aurisano.check.plot}
\alias{bmdh.aurisano.check.plot}
\title{Plot the different between the Aurisano and current BMDh values}
\usage{
bmdh.aurisano.check.plot(
  to.file = F,
  toxval.db = "res_toxval_v95",
  sys.date = "2024-02-23"
)
}
\arguments{
\item{to.file}{If TRUE, send the plot to a file}

\item{toxval.db}{Database version}

\item{sys.date}{The date of the database export}
}
\description{
`bmdh.per.study` plots the different between the Aurisano and current BMDh values
}
