% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdh.per.chemical.R
\name{bmdh.per.chemical}
\alias{bmdh.per.chemical}
\title{Calculate BMDh values one per chemical}
\usage{
bmdh.per.chemical(
  toxval.db = "res_toxval_v95",
  sys.date = "2024-02-23",
  regulatory.sources = c("IRIS", "PPRTV (CPHEA)", "ATSDR MRLs 2022", "ATSDR PFAS 2021",
    "EPA OPP", "HEAST")
)
}
\arguments{
\item{toxval.db}{Database version}

\item{sys.date}{The date of the database export}

\item{regulatory.sources}{This is the list of sources that will be used to select the
optimal quantile  to use for selecting the final chemical-level BMDh.}
}
\value{
Write a file with the results: toxval_PODs_for_BMDh chemical level {toxval.db} {sys.date}.xlsx
}
\description{
`bmdh.per.study` Calculates one BMDh value per chemical. This is done by taking
various percentiles of the distribution of the BMDh values and building a table
with one column per percentile per chemical. The values are calibrated
against regulatory values. The list of high-quality, regulator sources is given
as one of the calling arguments.
}
