% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdh.percentile.plot.R
\name{bmdh.percentile.plot}
\alias{bmdh.percentile.plot}
\title{Plot the BMDs vs teh regulatory values for different percentiles}
\usage{
bmdh.percentile.plot(
  to.file = F,
  toxval.db = "res_toxval_v95",
  sys.date = "2024-02-23",
  minstudies = 10,
  cutoff.logsd = 2
)
}
\arguments{
\item{to.file}{If TRUE, send the plot to a file}

\item{toxval.db}{Database version}

\item{sys.date}{The date of the database export}

\item{minstudies}{- only chemicals with this minimum number of studies will be used in the calculation}

\item{cutoff.logsd}{Only chemicals with their log SD of BMDh values will be used in the calculation}
}
\value{
Write a file with the results: toxval_PODs_for_BMDh chemical level {toxval.db} {sys.date}.xlsx
}
\description{
`bmdh.percentile.plot` Helps determine the optimal percentile. The output file shows the
fit statistics for different percentiles, and one should select the one with the lowest RMSE and highest R2.
}
