% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contains.R
\name{contains}
\alias{contains}
\title{Find out if one string contains another}
\usage{
contains(x, query, verbose = F)
}
\arguments{
\item{x}{The string to be searched in}

\item{query}{the second string}

\item{verbose}{if TRUE, the two strings are printed}
}
\value{
if x contains query, return TRUE, FALSE otherwise
}
\description{
Find out if one string contains another
}
