% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{driver}
\alias{driver}
\title{Run all of the calculations to go from database export to calcualtion of final BMDh values}
\usage{
driver(toxval.db = "res_toxval_v95", sys.date = "2024-02-23", user, password)
}
\arguments{
\item{toxval.db}{Database version}

\item{sys.date}{The date of the database export}

\item{user}{The username for the MySQL database. The database instance is
hard-coded in the function setDBConn().}

\item{password}{The user's MySQL database password.}
}
\description{
`driver` Run all of the calculations to go from database export to calcualtion of final BMDh values
}
