% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.for.bmdh.R
\name{filter.for.bmdh}
\alias{filter.for.bmdh}
\title{Filter the exported records for redundancy}
\usage{
filter.for.bmdh(toxval.db = "res_toxval_v95", sys.date = "2024-02-23")
}
\arguments{
\item{toxval.db}{Database version}

\item{sys.date}{The date of the export}
}
\value{
Write a file with the filtered results:ToxValDB for BMDh filtered {toxval.db} {sys.date}.xlsx
}
\description{
`filter.for.bmdh` Filters redundant rows in the raw database export. There are
two kinds of redundancy. The first filters extra reference rows from the record_source table.
The main data is in the toxval tables, and references are linked through the toxval_id to
the record_source table. During the curation process, these references get
cleaned and hence repeated, so a single (and the final or best) is selected.
Then, there are redundancies in the toxval table itself, likely caused by upstream
processing issues. These will be solved there, but the filtering
in this function takes care of this issue
for the moment.
}
