% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printCurrentFunction.R
\name{printCurrentFunction}
\alias{printCurrentFunction}
\title{Print the name of the current function}
\usage{
printCurrentFunction(comment.string = NA)
}
\arguments{
\item{comment.string}{An optinal string to be printed}
}
\description{
Print the name of the current function
}
