% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInsert.R
\name{runInsert}
\alias{runInsert}
\title{Insert a record into a database. if auto.increment=TRUE, return the auto incremented
primary key of the record. otherwise, return -1}
\usage{
runInsert(query, db, do.halt = F, verbose = F, auto.increment.id = F)
}
\arguments{
\item{query}{a properly formatted SQL query as a string}

\item{db}{the name of the database}

\item{do.halt}{if TRUE, halt on errors or warnings}

\item{verbose}{if TRUE, print diagnostic information}

\item{auto.increment}{if TRUE, add the auto increment primary key even if not part of the query}
}
\value{
Returns the database table auto incremented primary key ID
}
\description{
Insert a record into a database. if auto.increment=TRUE, return the auto incremented
primary key of the record. otherwise, return -1
}
