% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runInsertTable.R
\name{runInsertTable}
\alias{runInsertTable}
\title{Inserts multiple rows into a database table}
\usage{
runInsertTable(mat, table, db, do.halt = T, verbose = F, get.id = T)
}
\arguments{
\item{mat}{data frame containing the data, with the column names corresponding}

\item{table}{name of the database table to which data will be inserted}

\item{db}{the name of the database}

\item{do.halt}{if TRUE, halt on errors or warnings}

\item{verbose}{if TRUE, print diagnostic information}
}
\description{
Inserts multiple rows into a database table
}
