% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toxval.redundancies.R
\name{toxval.redundancies}
\alias{toxval.redundancies}
\title{Detect potential redundancies in ToxValDB.}
\usage{
toxval.redundancies(toxval.db = "res_toxval_v95", user = "rjudson", password)
}
\arguments{
\item{toxval.db}{Database version}

\item{user}{The username for the MySQL database. The database instance is
hard-coded in the function setDBConn().}

\item{password}{The user's MySQL database password.}
}
\value{
Write a file with the results
}
\description{
Detect potential redundancies in ToxValDB.
}
