% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmdh.aurisano.check.plot.R
\name{bmdh.aurisano.check.plot}
\alias{bmdh.aurisano.check.plot}
\title{Plot the different between the Aurisano and current BMDh values}
\usage{
bmdh.aurisano.check.plot(
  to.file = F,
  toxval.db = "res_toxval_v95",
  sys.date = "2024-01-04"
)
}
\arguments{
\item{dir}{The directory where the lists are stored}
}
\description{
Plot the different between the Aurisano and current BMDh values
}
